Add fail safety mechanism to config-set: --try-once
Sometimes wrong config settings may render device unusable, to the point where it goes into a state where it needs to be reflashed.

This change introduce a "try this config once" mechanism: when loading
configuration during boot, mos (the firmware) will check for config file with a `.try` suffix first:

 * conf0.json.try, then conf0.json
 * conf1.json.try, then conf1.json
 * ...
 * conf9.json.try, then conf9.json

If `.try` file exists, it will be loaded instead and deleted immediately after that.

This way, if device reboots for any reason, next boot will revert to the previous configuration.

`Config.Save` RPC gets a boolean `try_once` key that causes it to write
`conf9.json.try` instead of `conf9.json` (there is currently no interface to write other config levels).

So, let's say we have a `risky_setting` that we want to try but are not sure is safe.

```
 $ mos config-set --try-once risky_setting=true
Getting configuration
Setting new configuration...
Saving and rebooting...
Note: --try-once is set, config is valid for one reboot only
```

At this point device reboots. If something goes wrong and causes it to reboot, config is reverted and `risky_setting` is back to `false`.

If, however everything is fine, the device stays up. However, since `.try` file has been deleted, it will lose the setting on next reboot.
If we are satisfied with the configuration, we need to set it again without the flag.

```
 $ mos config-set --no-reboot risky_setting=true
Getting configuration
Setting new configuration...
Saving...
```

We used `--no-reboot` to avoid the reboot, we only want to commit the config.

CL: Add fail safety mechanism to config-set: --try-once

PUBLISHED_FROM=cbad6ff4ca044906b636e22618289769f034acc9